refactor: Update import paths and improve async function in get-magnet-health
