+++ src/main/events/catalogue/repacks/get-magnet-health.ts
@@ -1,5 +1,5 @@
-import { webTorrentData } from "@main/services/web-torrent-health"
-import { registerEvent } from "../register-event"
+import { webTorrentData } from "@main/services/web-torrent-data"
+import { registerEvent } from "../../register-event"
 
 const getMagnetHealth = async (_event: Electron.IpcMainInvokeEvent, magnet: string) => {
    return webTorrentData.getSeedersAndPeers(magnet)
